Fixed the Bucket code, some safety issues, and a couple small fixes
The in-place construction and destruction wasn't actually correct, since
the arrays called the destructors of each of their members upon Bucket
destruction, which would result in a double destruction.

Instead, we use arrays of aligned_storage objects, so that the
array destructor does nothing. This means we don't need to use custom
allocators, so we replaced the bucket pointer with a vector.

Also replaced arrays with std::array, and a few pointers with
unique_ptrs.

Also started using unique_lock in some places.