Re-did the operator[]
Made several major changes to the original implementation of operator[]

1. No longer do we insert a default value if the user doesn't do
anything with their reference. This is more similar to how languages
like Python work, and early on we decided we like this behavior better
than the C++ behavior. Also, since our reference is going to deviate
from STL behavior significantly anyways, we shouldn't be too concerned
with emulating it.

2. This means we don't have to worry as much about extending reference
lifetimes beyond what they are supposed to be, since multiple copies
won't have an effect on the table when they are destroyed. There is
still the issue of extending the reference beyond the lifetime of the
map it points to, but deleting the default reference constructor makes
that very hard to do, and also the STL doesn't worry about that either,
so it's probably not a big deal.

3. We can also support the more natural syntax "cuckoo_map[key] =
cuckoo_map2[key2]" by overloading the operator= for a "const reference&"
parameter to simply cast the argument to a mapped_type and use the other
version of operator=.

Also changed the test and moved it from test_iterator to a new file
called test_bracket_operator, since the tests didn't really fit anywhere
else.