Let the update function mutate the value argument
The original update function accepts a const value argument and must
return a new one that is copied back into the bucket. This can be
wasteful for mutable values like a string or an inner hash table.

We modified the update function specification to accept a mutable
reference to the value and do whatever with it. *NOTE* This is a
breaking change, since existing update functions need to be changed to
modify their argument and return nothing rather than return a new value.