Added support for inserting non-copy-constructable values into the table
Before, we were always copying values when you inserted something into
the table. This won't work if you wanted to use something like a
unique_ptr as your value, since you cannot copy a unique_ptr.

Now you can insert move-constructable items into the table, since
we forward the lvalue/rvalue-ness of the value you pass in with
std::forward.

This change also involved invalidating a number of methods that rely on
having copy-assignable values, like find, update, etc using SFINAE. As
an alternative to find, one can use update_fn with a lambda that reads
the value and does something with it.