Replaced iterator implementation with locked_table
Once the table is locked, we should be able to spawn and as many
STL-style iterators on the table as we want. The original iterator
implementation only allowed one iterator on the table, since the
iterator owns all the locks. But now users can create a "locked_table"
object, which owns all the locks and can generate iterators. The
iterators will die when the locked_table releases all its locks.