Made it easier to use different tables in the benchmarks
The test_utils.cc file now uses wrapper classes for the read_thread and
insert_thread methods.

So if somebody wanted to run the benchmarks with TBB, for instance, they
could add a partial template specialization of the read_thread and
insert_thread classes using a TBB table, replace the tables in the
benchmark files with TBB, and recompile (with the correct include/linker
flags as well). This approach should apply to any other table to be tested.