Re-did build process a bit
Mainly introduced cityhash as a sub-dependency, rather than trying to
compile in the necessary parts of it. This simplified the configure.ac
and Makefile.am scripts in a few directories.

Also re-named the libcuckoo directory to src (as it should be). This
means that users should run `make install` BEFORE compiling tests and
examples

Also replaced test_util.cc with test_util.hh, and included it in the
Makefile.am, so that it gets bundled when we run 'make dist'

Updated the README