Re-did the test setup, set up for actual unit tests
Separated the tests into unit-tests, stress-tests, and benchmarks. This
involved making the non-performance and non-stress type tests into
actual unit-tests, with the CATCH unit test framework.

Also changed the include paths in the tests and examples to be relative
to the project, not depend on the installation. Also, tests and examples
all compile with 'make all', and all tests are run with 'make check'.

More unit tests to come