Fixed constructor test
Our existing bound was too low, so we were still throwing on the first
allocation.

Instead, we break encapsulation a bit by adding a friend class for unit
testing, which knows the size of the TableInfo struct, so we can
allocate just enough for the first one. Then we fail on the second.