Allow freeing old TableInfo during resize
During cuckoo_expand_simple, we were never freeing the memory for the
old TableInfo, because the hazard pointer for the thread remained set to
the old table info. We can change it to point to the new table info,
since the current thread will not be using the old table info anymore.
This will allow us to delete the old table info, if no other thread is
looking at it during the expansion. This was tested manually.