Re-did lock container for locking whole table
Before, we would store a reference to the TableInfo as well as a dummy
container that would free the locks on the TableInfo when destroyed, but
these can be combined into a pointer container that frees the locks when
destroyed.

Integrating this with the iterator code simplified it as well.