Added minimum load factor to table
Now by default, we maintain a minimum load factor for the table, that
will be checked any time we try to automatically expand. If the load
factor drops below this minimum threshold, then we throw an exception.
This should prevent expanding forever in certain pathological cases such
as when a bad hash function or adversarial workload causes too many keys
to have the same hash value.