Added a maximum hash power argument to the table
It lets you cap the hash power the table can expand to, in any cases.
This is useful if people know they want a fixed sized data set and do
not want to expand beyond it. Rather than having to check the power
every time they expand, they can use this cap to bound the amount of
memory and (roughly) the number of elements the table can store.

Also cleaned up some tests and documentation.