Incorporated partial key into alt_index
Rather than using the full hash key as a tag for the alt_index function,
it's better to use the partial key (hashed to a 64-bit number), since
that avoids the overhead of re-hashing the key to get the alternate
bucket, when doing a cuckoopath_move. This also means that alt_index,
rather than using the next n bits of the hash key, will use the same tag
for the XOR.

This means that instead of getting rid of partial keys altogether for
simple types, we always store them, but ignore them when we're looking
for a key in a bucket if it's simple.