Added tests for bucket hashing
We want to test various properties of index_hash and alt_index that are
necessary for the correct and efficient functioning of the table

- You can recover the original bucket with alt_index if you pass in the
  alternate bucket
- The alternate bucket is pretty much always different from the index
  bucket
- Increasing the hashpower only grows the index_hash and alt_index by
  their top bits, leaving the bottom bits unchanged.

This found a bug in alt_index, where we were adding 1 to ensure the tag
is nonzero, which would fail when the partial was -1. Instead, we zero
out the first bit and add 1.