Fixed some tests
The partial_t typedef needs to be public for linux compilers.

Also, with a terrible integer hash function (like std::hash which just
returns the integer itself), the partial key ends up being the same for
all small keys (0 when the upper bits are 0). So we created default hash
class, which uses something different than std::hash for integral types,
and falls back to std::hash for everything else. It is intended to be
usable without any other dependencies, like CityHash, so we avoid
including CityHash.