Got rid of TableInfo and hazard pointers
The reason we had to store all the table data in one TableInfo pointer
and had hazard pointers was that during resizing, we would change to a
larger set of buckets, but we couldn't delete the old ones until all the
other threads were done using it. The way they would check that the
buckets changed was by checking if the global table_info pointer was
different from the one they held. The hazard pointers let us hold onto
an old version of the table until we were done using it, after which it
could be safely deleted.

We can get rid of all of this if we notice that the existing operations
are only invalidated if the hashpower changes. If the table is shrunk
and then expanded back to the original hashpower, other operations are
still valid, because the buckets that their hash key hashes to remain
the same. Thus we can invalidate operations when taking a lock, by
checking that the hashpower after taking the locks is the same as what
it was before.

If we impose the additional invariant that the hashpower and bucket
memory can only be changed when a thread holds all the locks (like it
would during resize), then we know that if a thread has a lock, then the
hashpower or bucket memory cannot be changed until the lock is released.
This lets us correctly operate with a certain hashpower after taking a
lock, until the lock is released.

Since the TableInfo and hazard pointer mechanisms ended up being fairly
low-cost, this change might not have a large performance implication
immediately, but it simplifies the table design a lot, and allows for
potentially faster resizing mechanisms.