Allow non-copyable keys and values properly
Implemented perfect forwarding for keys and values in the table. This
involved making moving data around buckets possible. It also prevents us
from copying keys inside the table, which means the old method of
verifying that a cuckoo spot is still valid in cuckoopath_move, which
compared the keys for equality, is not valid. Luckily, as long as the
hash values match, the cuckoo operation is still valid, because equal
hash keys will hash to the same buckets, so we only need to store the
hash key.