Properly construct key and value arguments inside bucket
Before, we were trying to forward all the arguments (single key argument
and multiple value arguments) to the pair constructor. But what we want
is to separately construct the key and the values. So now setKV takes in
the key separately from the values and constructs them piecewise as
separate tuples. This preserves forwarding semantics.