Request structure alignment in a portable way.
libcuckoo uses a GCC-specific attribute to request structure alignment. This
commit replaces this with a new LIBCUCKOO_ALIGNAS macro which uses the
GCC-specific attribute with GCC, but the C++-standard alignas() keyword
elsewhere.

This commit also adds (and uses) the macro LIBCUCKOO_SQUELCH_PADDING_WARNING to
suppress a warning from MSVC about the alignment request resulting in a padded
structure.