Added support for heterogeneous compare
C++14 added support for heterogeneous comparison in containers, which
allows, for example, running `find` inside an std::map<std::string>
container with a `const char*` key, and avoiding the construction of an
std::string out of the key argument, since std::string has an overload
for comparing against `const char*`.

This commit adds support for passing heterogeneous equality and hash
functions to the table.