Added test for exceptions in user code
The table preserves the strong exception guarantee, which means if
exceptions are thrown while running user code in one thread, the
exception will propagate, but the table will still be usable in other
threads.

We had to make some fixes for handling exceptions in threads spawned by
the table, as well as not calling cuckoo_fast_double unless the move
constructor is noexcept.