Change total ops to be a percentage of initial capacity
This way, when we do a pure insert benchmark, we won't be filling the
table to 100% capacity, which will almost certainly trigger an
expansion. The default is 90, so the number of operations we do is 90%
of the table capacity. For more stability (or to insert beyond the
capacity of the table), this can be changed to a percentage greater than
100.