Make key gen use smaller numbers
Our key generation was generating keys that only differed between
threads in the upper bits. With std::hash for integers being the
identity function, this turns out to be really bad, since all the keys
from different threads will hash to the same buckets.

We could use a different hash function, but for speed and simplicity, we
decided to change the key generation function to generate multiples of
the total number of threads, offset by the thread id.