Fixed a couple things with the cityhash build
Rather than running configure in the source tree, run it in a separate
'cityhash-build' directory, rooted in the binary tree. Also, include the
current directory when building cityhash, since some of the includes are
of the form #include <city.h>, which means the current directory needs
to be an include path.