API cleanups part 1
Started to organize the code into public-first and private-later. Added richer
documentation and grouping to the public types.

Removed the pseudo-reference class. I think it won't fit well with the upcoming
API changes, where the locked_table will look a lot like the STL hash table, and
we have a fancier upsert function. This means removing the
`test_bracket_operator` unit test.

We also changed the slot_per_bucket member to a function, which is more in line
with the other property accessors, which required tiny modifications to a lot of
unit tests.

Behaviorally, we clarified the lifetime behavior of objects being inserted. The
key must be constructed immediately, because we need to hash and compare with
it. But we also want to allow non-copyable keys, so we immediately construct the
key template with a move constructor, then move it into the table if the insert
succeeds, or destroy it if the insert fails.