Added C port of libcuckoo
The source is located in a new subfolder called 'libcuckoo-c'. It is
structured as a template library, wherein a user can define macros
specifying the name of table, type of key, and type of value, then
include the proper header and implementation files to fill in the table
definition. The header and implementation can then be compiled as a
C-linkable library.

There is one example in the examples directory "hellohash_c.c" which
uses a custom int->string table. The unit test "test_c_interface.cc"
contains good demonstrations of every feature of the port.

There were a couple of minor  C++ cleanups that took place. 1) Fixed the
assertions about locking not to throw if we are in locked_table mode and
2) remove an extraneous inclusion in hellohash.cc