Added an example of using untyped blobs in the C interface
Created a blob->blob table with separate key and mapped blob types.
Replaced hellohash_c with c_hash, which combines usage of the int_str
table and the blob_blob table in order to showcase using multiple tables
in the same file. The blob_blob example uses the locked_table and
iterator interfaces.

Added explanatory comments to the C example files to explain a bit more
about what's going on.