Split libcuckoo-c into interface and implementation libraries
This is needed because the interface (header template) needs to be propagated
publicly in a target like `int_int_table`, but the implementation
(implementation template) nedes to be linked privately, so that we don't
unnecessarily propagate the libcuckoo C++ library and its compiler options when
we link with `int_int_table`. There's no good way in CMake to specify that a
library should propagate options for direct linkage but not second-level
linkage, so we split into two libraries.

Noticed this problem when compiling `c_hash` with gcc, which complained that we
were adding the flag `std=gnu++11`, which only applies to a C++ compiler and
was being included because we were propagating the `libcuckoo` interface
compile options.