Re-factored storage into libcuckoo_bucket_container.hh
The logic to manage allocators and uninitialized storage ended up growing quite
complicated, so it made sense to re-factor it into a separate container. Took a
lot of care to use allocators correctly, and allocate and free memory with the
correct allocators.

Significant changes include a change to the value_type, from std::pair<const
Key, T> to an aggregate wrapper type which just includes references to key and
mapped data. This isn't a breaking change per se, in that it preserves the
value_type interface, but the type is different. This change was necessary
because there is no way to move a std::pair<const Key, T>, due to the const
key, without invoking undefined behavior. So both the key and value must be
stored as non-const data, and the key only exposed as const.

This has the advantage of 1) allowing us to store all the keys in a bucket
contiguously, and all the mapped values contiguously, which is better for
alignment. It also opens up the possibility of serializing the BucketContainer
data.

The new bucket container also supports copying and moving, which opens up the
possibility for making the entire cuckoohash_map copyable and movable, though
the concurrency situation will have to be taken into consideration.