Made lock array allocator aware
Since spinlock is a custom-defined POD type, it doesn't need a special
allocator to construct it, and we don't really need a generic lazy_array just
for locks. So we specialize the lazy array just for spinlocks, and get rid of
some extraneous exception-handling code. We added in allocator-aware moves,
copies, and assignments, so that, in the future, we might be able to do that
for the entire hash table. The catch is that, when copying, only the
elem_counter is changed, not the state of the lock, since that generally
doesn't make sense.

Furthermore, we fixed a few bugs regarding reserve calculation, constants, and
updating the element counters in the lock array.

We also made the LOCK_GRANULARITY a template paramater, since it is useful for
testing, and will allow users to have different tables with different
granularities (useful for nesting tables.