Added layout compatibility check to bucket container
Since we're interoperating between std::pair<const Key, T> and std::pair<Key,
T>, we need to make sure this is safe. We check that both classes are
standard_layout and that the members have the same struct offsets. If this is
true, then the aligned_storage should work correctly for both types.