Small cleanups
Removed the `static_assert` added earlier, and replaced with a warning about
user-specialized instances of `std::pair`. Ultimately, it's unlikely there will
be a portable and exactly correct way to specify the necessary requirements for
`std::pair`, and so we resort to a textual warning. This is also done by C++17
in their node handle documentation.

Also fixed the case statement in `cuckoo_insert_loop`. It was technically
correct, it relied on a case statement falling through, which is somewhat bad
style, and was causing a warning on GCC 7.1. Changed to make each case
independent.