Fixed allocator-handling in lock array and bucket container. Also decreased lock array class size.
According to the AllocatorAwareContainer specification, an allocator rebound to
a different type must compare equal to the original container. Furthermore, the
stated `allocator_type` must be bound to allocate instances of `value_type`.
Finally, the allocator bound type only matters when allocating memory, but any
allocator type can be used for construction.

This greatly simplifies the allocator handling logic in these containers. Now,
we have one actual allocator, which is for instances of `value_type`. Other
allocators are simply copies of this original allocator, which are only
maintained so a copy doesn't need to be made every time we need to allocate
memory for a different type. These copies can always just be copies of the
original allocator, so when we copy/move/swap, we can just copy them again from
the original, rather than implement some complex logic for conditionally
copy/move/swapping all of our allocators.

Another big change was to replace the std::array<spinlock*, NUM_SEGMENTS> array
in libcuckoo_lock_array with spinlock**. This was important because if we had
LIBCUCKOO_LOCK_ARRAY_GRANULARITY set to 16, our lock_array class would have
size ~2^16 (NUM_SEGMENTS) * 2^3 (pointer size), which is way too big and could
blow up the stack. It remains to be seen whether this causes a performance
degradation.