Resize locks array during expansion Got rid of expansion_lock.
expansion_lock is unnecessary.

During expansion, we take all the old locks, and maybe replace it with a larger
locks array. We must make sure the lock resize is done before setting the
hashpower, and that the new locks are all held before they are made visible.
If these two things are not done, then it is possible for other threads to run
table operations, either by grabbing the new hashpower and locking the old
locks (if we don't do the first thing), or by grabbing the new hashpower and
locking the new locks before moving is complete (if we don't do the second
thing).

Also make sure locked_table grabs the correct set of locks, in case it is
conflicting with an expansion.