Globalized locked table mode
There was an issue with the original code, where if a locked table method
initiated a resize (through an insert that needed to expand, or through an
explicit rehash or reserve), the locked table wouldn't grab hold of the new
locks, if a new locks container was allocated. We needed some way of holding
onto the locks if we're in locked table mode, and letting them go otherwise.

To fix this, we store a member variable, which keeps track of which thread has
locked the table (if any). This thread keeps track of an iterator into
`all_locks_`, which marks the position of the locks it has. It is presumed this
thread is responsible for freeing all locks containers from this iterator until
the end, whenever it is done.

Now, since this state is a member variable, we don't need an explicit LOCK_T to tell
us whether to take locks or not, we simply inspect this variable.

Also includes some random cleanups.