Got rid of run-time locked table status
It caused a slight performance degradation in reads, and it's ultimately
unnecessary. The lock containers (LockManager, TwoBuckets, AllUnlocker) can be
agnostic of locking mode, but we have to re-introduce the templating everywhere
else.

Rather than tying it to the locking type, which isn't totally clear, we
re-named it to TABLE_MODE, and differentiate between `locked_table_mode` (locks
already taken) and `normal_mode` (no locks taken).