Fix a type truncation error from MSVC 2015 in cuckoohash_map::size().
1>cuckoohash_map.hh(344): warning C4244: 'return': conversion from 'cuckoohash_map<...>::counter_type' to 'unsigned int', possible loss of data

Although it might be possible in theory for the sum of each buckets' element
count to overshoot 2^32 in a 32-bit build such that cuckoohash_map::size()
returns a truncated count, in practice I doubt enough elements can be stored
in the table to hit this problem. This commit just adds a cast to silence the
warning.