Silence bogus deadcode warning from MSVC 2015.
1>cuckoohash_map.hh(1581): warning C4702: unreachable code

This code (exception handling around a call to move_buckets()) is only
unreachable if the constructors for our keys and values cannot throw. This will
not be the case for all instantiations of this cuckoohash_map. So this commit
adds the macros LIBCUCKOO_SQUELCH_DEADCODE_WARNING_BEGIN /
LIBCUCKOO_SQUELCH_DEADCODE_WARNING_END and uses them to disable MSVC's deadcode
warning for this sequence of code.