Improvements to slot_search
There were a few apparent inefficiencies with the slot_search function:
  - We were re-locking the bucket for every individual slot we inspected, but
    it seems we can just lock the bucket in the outer loop.
  - We have a MAX_BFS_PATH_LEN, but limit the size of the queue by a separate
    constant. It might be easier to just adjust MAX_BFS_PATH_LEN to change the
    depth of the cuckoo search, and complete the search all the way. This ensures
    we always consider the entire search space in a slot_search.

In benchmarking, we find the maximum load capacity of the table increases, from
roughly 89% with integer keys to roughly 96% with integer keys. There is a
slight decline in performance from slightly longer cuckooing operations, but
it's not very significant.  This tradeoff can be tweaked if necessary by
changing MAX_BFS_PATH_LEN.