Fix friend declarations that refer to templates, not classes.
This commit addresses a couple of PW.BAD_FRIEND_DECL parse warnings thrown up by
Coverity.

Apparently a friend declaration that refers to a template (as opposed to a type)
is not legal C++. Who knew? Neither GCC nor MSVC seem to care. Anyhow, this
commit corrects the error by specifying all the necessary template parameters to
make the declarations refer to types.