Wrap everything in libcuckoo namespace.
This is a breaking change. Users will have to add the namespace prefix
`libcuckoo::` to types and constants, or bring them into the global namespace
with something like `using libcuckoo::cuckoohash_map`.

This also allows us to get rid of the `std::swap` specialization, which is not
allowed by C++. Instead, we provide a specialization within the `libcuckoo`
namespace, which can be found using argument-dependent lookup.