Add cuckoohash_map::find_or_insert_fn
This new API makes it trivial to leverage libcuckoo for string
interning or similar areas, where constructing a value is potentially
expensive or requires further book keeping:

```
    std::shared_mutex mutex;
    std::vector<std::string> indexToString;
    libcuckoo::cuckoohash_map<std::string, std::size_t> stringToIndex;

    auto stringForIndex = [&](std::size_t index)
    {
        auto lock = std::shared_lock(mutex);
        return indexToString[index];
    };

    auto indexForString = [&](const std::string &string)
    {
        return stringToIndex.find_or_insert_fn(string, [&]() {
            auto lock = std::unique_lock(mutex);
            auto index = indexToString.size();
            indexToString.push_back(string);
            return index;
        });
    };
```

Note how we only need to acquire the expensive unique_lock when we
are inserting a new string key value. In the hopefully common case
where the string was already inserted previously, we can rely on
the shared libcuckoo locking to reduce lock contention.

I do not see any way to leverage the existing libcuckoo API in a
similar manner - but maybe I'm missing something? Note how I cannot
use something like the following, as it would contain race conditions
for the size call and subsequent insertion into the indexToString
vector:

```
    // problem #1: multiple threads could read the same index value here
    auto index = stringToIndex.size();
    auto wasInserted = stringToIndex.upsert(string, [&index](const auto &existingIndex) {
        index = existingIndex;
    }, index);
    if (wasInserted) {
        // problem #2: order in which indexToString is updated is undefined
        auto lock = std::unique_lock(mutex);
        indexToString.push_back(string);
   }
```