Fix segfault.
It was possible to produce a segfault with `universal_benchmark` when
running a script like the following (on a `t3.2xlarge` ec2 machine with
amazon linux):

```
for i in {1..100}; do "./tests/universal-benchmark/universal_benchmark" "--inserts" "100" "--initial-capacity" "1" "--total-ops" "131072000" "--num-threads" "16" || break; done
```

From adding print statements, we observed that some thread (call it
thread X) would be trying to lock the table, when other threads would
manage to resize the table twice underneath the nose of thread X. This
meant that both the bucket memory thread X was looking at was invalid,
and furthermore that the locks array thread X was looking at wasn't
being locked. Then thread X would acquire its lock, find its hashpower
to be the same as always, and then segfault trying to read the old
bucket memory.

What didn't add up is how thread X was reading the current hashpower to
be the same as before the two resizes completed. Without understanding
fully, my educated guess is that since the buckets_ array is a standard
member variable, when we replace it as `buckets_ = buckets_t(new_hp,
...)`, this replacement isn't fully atomic, so somehow the old thread
manages to read the old value of the hashpower, perhaps from an old
piece of memory. Maybe the address of the old buckets container was
stored in a register for that thread or something.

Regardless, the way we fix it is to introduce a separate atomic counter
called `resize_counter`, which we use to signal each time the bucket
container is resized. By synchronizing on the atomic variable, we are
able to detect the resize reliable across threads at the minimal cost of
reading one extra atomic variable.